apps/api/pyproject.toml
[project]
name = "code9-api"
version = "0.1.0"
description = "Code9 Analytics — FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111,<0.120",
  "uvicorn[standard]>=0.30,<0.35",
  "pydantic>=2.7,<3",
  "pydantic-settings>=2.3,<3",
  "sqlalchemy[asyncio]>=2.0,<2.1",
  "alembic>=1.13,<2",
  "asyncpg>=0.29,<0.31",
  "psycopg2-binary>=2.9,<3",
  "python-jose[cryptography]>=3.3,<4",
  "argon2-cffi>=23.1,<24",
  "cryptography>=42,<44",
  "redis>=5,<6",
  "rq>=1.16,<2",
  "httpx>=0.27,<0.29",
  "python-multipart>=0.0.9,<0.1",
  "email-validator>=2.1,<3",
  "orjson>=3.10,<4",
  "structlog>=24,<25"
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5,<0.9",
  "mypy>=1.10,<2",
  "pytest>=8,<9",
  "pytest-asyncio>=0.23,<0.25",
  "httpx>=0.27,<0.29"
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
